feat(core): split range facet into core (#1245)
* feat(core): split range facet into core

[COM-1285]
