+++ packages/headless/src/controllers/core/facets/range-facet/date-facet/headless-core-date-facet.test.ts
@@ -1,30 +1,27 @@
 import {
   DateFacet,
-  buildDateFacet,
+  buildCoreDateFacet,
   DateFacetOptions,
   buildDateRange,
-} from './headless-date-facet';
+} from './headless-core-date-facet';
 import {
   MockSearchEngine,
   buildMockSearchAppEngine,
-} from '../../../../test/mock-engine';
-import {createMockState} from '../../../../test/mock-state';
-import {executeSearch} from '../../../../features/search/search-actions';
+} from '../../../../../test/mock-engine';
+import {createMockState} from '../../../../../test/mock-state';
 import {
   deselectAllDateFacetValues,
   registerDateFacet,
   toggleSelectDateFacetValue,
-} from '../../../../features/facets/range-facets/date-facet-set/date-facet-actions';
-import {buildMockDateFacetValue} from '../../../../test/mock-date-facet-value';
-import {buildMockDateFacetResponse} from '../../../../test/mock-date-facet-response';
-import {SearchAppState} from '../../../../state/search-app-state';
-import * as FacetIdDeterminor from '../../../core/facets/_common/facet-id-determinor';
-import {buildMockDateFacetRequest} from '../../../../test/mock-date-facet-request';
-import {configuration, dateFacetSet, search} from '../../../../app/reducers';
-import {updateFacetOptions} from '../../../../features/facet-options/facet-options-actions';
-import {DateFacetValue} from '../../../../features/facets/range-facets/date-facet-set/interfaces/response';
-import {deselectAllFacetValues} from '../../../../features/facets/facet-set/facet-set-actions';
-import {updateRangeFacetSortCriterion} from '../../../../features/facets/range-facets/generic/range-facet-actions';
+} from '../../../../../features/facets/range-facets/date-facet-set/date-facet-actions';
+import {buildMockDateFacetValue} from '../../../../../test/mock-date-facet-value';
+import {buildMockDateFacetResponse} from '../../../../../test/mock-date-facet-response';
+import {SearchAppState} from '../../../../../state/search-app-state';
+import * as FacetIdDeterminor from '../../_common/facet-id-determinor';
+import {buildMockDateFacetRequest} from '../../../../../test/mock-date-facet-request';
+import {configuration, dateFacetSet, search} from '../../../../../app/reducers';
+import {updateFacetOptions} from '../../../../../features/facet-options/facet-options-actions';
+import {DateFacetValue} from '../../../../../features/facets/range-facets/date-facet-set/interfaces/response';
 
 describe('date facet', () => {
   const facetId = '1';
@@ -35,7 +32,7 @@ describe('date facet', () => {
 
   function initDateFacet() {
     engine = buildMockSearchAppEngine({state});
-    dateFacet = buildDateFacet(engine, {options});
+    dateFacet = buildCoreDateFacet(engine, {options});
   }
 
   beforeEach(() => {
@@ -99,65 +96,6 @@ describe('date facet', () => {
       const action = toggleSelectDateFacetValue({facetId, selection: value});
       expect(engine.actions).toContainEqual(action);
     });
-
-    it('dispatches a search', () => {
-      const value = buildMockDateFacetValue();
-      dateFacet.toggleSelect(value);
-
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-      expect(action).toBeTruthy();
-    });
-  });
-
-  describe('#deselectAll', () => {
-    beforeEach(() => dateFacet.deselectAll());
-
-    it('dispatches #deselectAllFacetValues with the facet id', () => {
-      expect(engine.actions).toContainEqual(deselectAllFacetValues(facetId));
-    });
-
-    it('dispatches a #updateFacetOptions action with #freezeFacetOrder true', () => {
-      expect(engine.actions).toContainEqual(
-        updateFacetOptions({freezeFacetOrder: true})
-      );
-    });
-
-    it('dispatches a search', () => {
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-
-      expect(engine.actions).toContainEqual(action);
-    });
-  });
-
-  describe('#sortBy', () => {
-    it('dispatches #updateRangeFacetSortCriterion', () => {
-      const criterion = 'descending';
-      dateFacet.sortBy(criterion);
-      const action = updateRangeFacetSortCriterion({facetId, criterion});
-
-      expect(engine.actions).toContainEqual(action);
-    });
-
-    it('dispatches a #updateFacetOptions action with #freezeFacetOrder true', () => {
-      dateFacet.sortBy('descending');
-
-      expect(engine.actions).toContainEqual(
-        updateFacetOptions({freezeFacetOrder: true})
-      );
-    });
-
-    it('dispatches a search', () => {
-      dateFacet.sortBy('descending');
-
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-      expect(action).toBeTruthy();
-    });
   });
 
   function testCommonToggleSingleSelect(facetValue: () => DateFacetValue) {
@@ -175,15 +113,6 @@ describe('date facet', () => {
       expect(engine.actions).toContainEqual(
         updateFacetOptions({freezeFacetOrder: true})
       );
-    });
-
-    it('dispatches a search', () => {
-      dateFacet.toggleSingleSelect(facetValue());
-
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-      expect(action).toBeTruthy();
     });
   }
