+++ packages/headless/src/controllers/core/facets/range-facet/date-facet/headless-core-date-filter.test.ts
@@ -1,27 +1,26 @@
-import {configuration, dateFacetSet, search} from '../../../../app/reducers';
-import {SearchAppState} from '../../../../state/search-app-state';
+import {configuration, dateFacetSet, search} from '../../../../../app/reducers';
+import {SearchAppState} from '../../../../../state/search-app-state';
 import {
   buildMockSearchAppEngine,
   createMockState,
   MockSearchEngine,
-} from '../../../../test';
-import {buildMockDateFacetRequest} from '../../../../test/mock-date-facet-request';
+} from '../../../../../test';
+import {buildMockDateFacetRequest} from '../../../../../test/mock-date-facet-request';
 import {
-  buildDateFilter,
+  buildCoreDateFilter,
   DateFilter,
   DateFilterInitialState,
   DateFilterOptions,
-} from './headless-date-filter';
-import * as FacetIdDeterminor from '../../../core/facets/_common/facet-id-determinor';
+} from './headless-core-date-filter';
+import {buildDateRange} from './date-range';
+import * as FacetIdDeterminor from '../../_common/facet-id-determinor';
 import {
   registerDateFacet,
   updateDateFacetValues,
-} from '../../../../features/facets/range-facets/date-facet-set/date-facet-actions';
-import {buildMockDateFacetValue} from '../../../../test/mock-date-facet-value';
-import {executeSearch} from '../../../../features/search/search-actions';
-import {updateFacetOptions} from '../../../../features/facet-options/facet-options-actions';
-import {buildMockDateFacetResponse} from '../../../../test/mock-date-facet-response';
-import {buildDateRange} from '../../../core/facets/range-facet/date-facet/date-range';
+} from '../../../../../features/facets/range-facets/date-facet-set/date-facet-actions';
+import {buildMockDateFacetValue} from '../../../../../test/mock-date-facet-value';
+import {updateFacetOptions} from '../../../../../features/facet-options/facet-options-actions';
+import {buildMockDateFacetResponse} from '../../../../../test/mock-date-facet-response';
 
 describe('date filter', () => {
   const facetId = '1';
@@ -33,7 +32,7 @@ describe('date filter', () => {
 
   function initDateFilter() {
     engine = buildMockSearchAppEngine({state});
-    dateFacet = buildDateFilter(engine, {options, initialState});
+    dateFacet = buildCoreDateFilter(engine, {options, initialState});
   }
 
   beforeEach(() => {
@@ -116,16 +115,6 @@ describe('date filter', () => {
       expect(engine.actions).toContainEqual(action);
     });
 
-    it('dispatches a search', () => {
-      const value = buildMockDateFacetValue();
-      dateFacet.setRange(value);
-
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-      expect(action).toBeTruthy();
-    });
-
     it('should return true when range is valid', () => {
       const value = buildMockDateFacetValue(
         buildDateRange({start: 1616592691000, end: 1616592691000})
@@ -154,14 +143,6 @@ describe('date filter', () => {
       expect(engine.actions).toContainEqual(
         updateFacetOptions({freezeFacetOrder: true})
       );
-    });
-
-    it('dispatches a search', () => {
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-
-      expect(engine.actions).toContainEqual(action);
     });
   });
