+++ packages/headless/src/controllers/core/facets/range-facet/headless-core-range-facet.test.ts
@@ -1,22 +1,21 @@
 import {
   MockSearchEngine,
   buildMockSearchAppEngine,
-} from '../../../test/mock-engine';
-import {createMockState} from '../../../test/mock-state';
-import {executeSearch} from '../../../features/search/search-actions';
-import {buildMockNumericFacetValue} from '../../../test/mock-numeric-facet-value';
-import {buildMockNumericFacetResponse} from '../../../test/mock-numeric-facet-response';
+} from '../../../../test/mock-engine';
+import {createMockState} from '../../../../test/mock-state';
+import {buildMockNumericFacetValue} from '../../../../test/mock-numeric-facet-value';
+import {buildMockNumericFacetResponse} from '../../../../test/mock-numeric-facet-response';
 import {
-  buildRangeFacet,
+  buildCoreRangeFacet,
   RangeFacet,
   RangeFacetProps,
-} from './headless-range-facet';
-import {updateRangeFacetSortCriterion} from '../../../features/facets/range-facets/generic/range-facet-actions';
-import {NumericFacetRequest} from '../../../features/facets/range-facets/numeric-facet-set/interfaces/request';
-import {buildMockNumericFacetRequest} from '../../../test/mock-numeric-facet-request';
-import {deselectAllFacetValues} from '../../../features/facets/facet-set/facet-set-actions';
-import {updateFacetOptions} from '../../../features/facet-options/facet-options-actions';
-import {SearchAppState} from '../../../state/search-app-state';
+} from './headless-core-range-facet';
+import {updateRangeFacetSortCriterion} from '../../../../features/facets/range-facets/generic/range-facet-actions';
+import {NumericFacetRequest} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/request';
+import {buildMockNumericFacetRequest} from '../../../../test/mock-numeric-facet-request';
+import {deselectAllFacetValues} from '../../../../features/facets/facet-set/facet-set-actions';
+import {updateFacetOptions} from '../../../../features/facet-options/facet-options-actions';
+import {SearchAppState} from '../../../../state/search-app-state';
 
 describe('range facet', () => {
   const facetId = '1';
@@ -27,7 +26,7 @@ describe('range facet', () => {
 
   function initRangeFacet() {
     engine = buildMockSearchAppEngine({state});
-    rangeFacet = buildRangeFacet(engine, props);
+    rangeFacet = buildCoreRangeFacet(engine, props);
   }
 
   beforeEach(() => {
@@ -79,14 +78,6 @@ describe('range facet', () => {
         updateFacetOptions({freezeFacetOrder: true})
       );
     });
-
-    it('dispatches a search', () => {
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-
-      expect(engine.actions).toContainEqual(action);
-    });
   });
 
   describe('#state.hasActiveValues', () => {
@@ -122,15 +113,6 @@ describe('range facet', () => {
       expect(engine.actions).toContainEqual(
         updateFacetOptions({freezeFacetOrder: true})
       );
-    });
-
-    it('dispatches a search', () => {
-      rangeFacet.sortBy('descending');
-
-      const action = engine.actions.find(
-        (a) => a.type === executeSearch.pending.type
-      );
-      expect(action).toBeTruthy();
     });
   });
