+++ packages/headless/src/features/facets/range-facets/numeric-facet-set/numeric-facet-controller-actions.ts
@@ -1,6 +1,5 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import {NumericFacetValue} from './interfaces/response';
-import {AsyncThunkSearchOptions} from '../../../../api/search/search-api-client';
 import {
   ConfigurationSection,
   NumericFacetSection,
@@ -11,6 +10,7 @@ import {facetIdDefinition} from '../../generic/facet-actions-validation';
 import {RecordValue} from '@coveo/bueno';
 import {numericFacetValueDefinition} from '../generic/range-facet-validate-payload';
 import {updateFacetOptions} from '../../../facet-options/facet-options-actions';
+import {AsyncThunkOptions} from '../../../../app/async-thunk-options';
 
 const definition = {
   facetId: facetIdDefinition,
@@ -30,7 +30,7 @@ export const executeToggleNumericFacetSelect = createAsyncThunk<
     facetId: string;
     selection: NumericFacetValue;
   },
-  AsyncThunkSearchOptions<ConfigurationSection & NumericFacetSection>
+  AsyncThunkOptions<ConfigurationSection & NumericFacetSection>
 >(
   executeToggleNumericFacetSelectType,
   (payload, {dispatch, extra: {validatePayload}}) => {
